+++ examples/chat/src/client/mod.rs
@@ -1,3 +1,4 @@
+pub mod chat;
 pub mod client;
 pub mod io;
 pub mod parser;
