+++ examples/chat/src/server/messages.rs
@@ -2,130 +2,14 @@ use core::fmt;
 ///! Definition of messages which can be produces by server.
 use std::time::SystemTime;
 
-use chrono::DateTime;
-use chrono::Local;
-use colored::Colorize;
-
+use chrono::{DateTime, Local};
 use dsbuild::Message;
 use serde::{Deserialize, Serialize};
 
+use super::chat_event::ChatEvent;
+
 /// Represents response from server to client request.
 pub type RequestResponse = Result<(), String>;
-
-/// Represents event in the chat.
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub enum ChatEventKind {
-    SentMessage(String), // Client sent message,
-    Connected(),         // Client connected to server,
-    Disconnected(),      // Client disconnected from server,
-    Created(),           // Client created chat.
-}
-
-/// Represents chat event born by request of some client.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ChatEvent {
-    chat: String,
-    client: String,
-    time: SystemTime,
-    kind: ChatEventKind,
-    seq: usize,
-}
-
-impl fmt::Display for ChatEvent {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let dt: DateTime<Local> = self.time.clone().into();
-
-        match &self.kind {
-            ChatEventKind::SentMessage(msg) => write!(
-                f,
-                "[{}]\t{} {} {}: {}",
-                dt.format("%Y-%m-%d %H:%M:%S").to_string().italic(),
-                self.client.bold().green(),
-                "->".green(),
-                self.chat.bold().green(),
-                msg.italic()
-            ),
-            ChatEventKind::Connected() => {
-                write!(
-                    f,
-                    "[{}]\t{} {} {}",
-                    dt.format("%Y-%m-%d %H:%M:%S").to_string().italic(),
-                    self.client.bold().green(),
-                    "connected to".green(),
-                    self.chat.bold().green()
-                )
-            }
-            ChatEventKind::Disconnected() => write!(
-                f,
-                "[{}]\t{} {} {}",
-                dt.format("%Y-%m-%d %H:%M:%S").to_string().italic(),
-                self.client.bold().green(),
-                "disconnected from".green(),
-                self.chat.bold().green()
-            ),
-            ChatEventKind::Created() => write!(
-                f,
-                "[{}]\t{} {} {}",
-                dt.format("%Y-%m-%d %H:%M:%S").to_string().italic(),
-                self.client.bold().green(),
-                "created".green(),
-                self.chat.bold().green()
-            ),
-        }
-    }
-}
-
-impl ChatEvent {
-    pub fn message_sent(chat: String, client: String, msg: String, seq: usize) -> Self {
-        Self {
-            chat,
-            client,
-            time: SystemTime::now(),
-            kind: ChatEventKind::SentMessage(msg),
-            seq,
-        }
-    }
-
-    pub fn client_connected(chat: String, client: String, seq: usize) -> Self {
-        Self {
-            chat,
-            client,
-            time: SystemTime::now(),
-            kind: ChatEventKind::Connected(),
-            seq,
-        }
-    }
-
-    pub fn client_disconnected(chat: String, client: String, seq: usize) -> Self {
-        Self {
-            chat,
-            client,
-            time: SystemTime::now(),
-            kind: ChatEventKind::Disconnected(),
-            seq,
-        }
-    }
-
-    pub fn chat_created(client: String, chat: String, seq: usize) -> Self {
-        Self {
-            chat,
-            client,
-            time: SystemTime::now(),
-            kind: ChatEventKind::Created(),
-            seq,
-        }
-    }
-
-    pub fn new_with_kind(chat: String, client: String, kind: ChatEventKind, seq: usize) -> Self {
-        Self {
-            chat,
-            client,
-            time: SystemTime::now(),
-            kind,
-            seq,
-        }
-    }
-}
 
 /// Represents messages from server to the client.
 ///
